timeit.py
QuickSort and MergeSort are both algorithms based on a process that repeately split a list of elements, in order to "isolate" each of the element to recombine theme
Mergesort repeated divides until one element, then compares and swaps each side as it merges two subarrays. Quicksort doesn't necessary split the list into half,  bacause it uses a pivot, iterates index and compares to pivot, swaps to left if less, right if more; repeats those steps to second side, then chooses a new pivot in one of the half’s, repeats the whole process.

MergeSort is a more stable sort than QuickSort and can be easily adapted to operate on linked lists and very large lists. But at the same time QuickSort is "Quicker" since it doesn't have to analyze additional new created lists that are create during the process.